requirements.txt
pyTelegramBotAPI
python-dotenv
requests
beautifulsoup4
Levenshtein
